添加获取学生作答功能

+++ src/backend/api/Test.go
@@ -147,11 +147,11 @@ func queryTestByID(db *gorm.DB, id int) []int {
 	return quesId
 }
 
-func GetGradeByTestIdAndQuestionId(db *gorm.DB, testId int, questionId int) (float64, error) {
+func GetGradeByTestIdAndQuestionId(db *gorm.DB, testId int, questionId int) (int, error) {
 	// 查询该题目的分数
 	var test Tests
 	if err := db.Table("tests").Where("id = ? AND question_id = ?", testId, questionId).Find(&test).Error; err != nil {
 		return 0, err
 	}
-	return test.Grade, nil
+	return int(test.Grade), nil
 }
